class 이름 수정, User class의 id와 관련class 수정
User의 id, long 타입에서 int로 수정
UserResponse의 id, long 타입에서 int로 수정
UserRepository extends, Long에서 Integer로 수정

Board -> BoardRepository 로 수정
UserAssignment -> UserAssignmentRepository 로 수정

+++ lms/src/main/java/net/skhu/lms/entity/User.java
@@ -12,7 +12,7 @@ public class User {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private int id;
 
 	@Column(unique = true, updatable = false, nullable = false)
 	private String userId;
